Improve efficiency for warrior hs/cleave, and use an aura so we can see queues and dequeues in the UI
