+++ sim/warrior/dps/rotation.go
@@ -188,15 +188,7 @@ func (war *DpsWarrior) tryMaintainDebuffs(sim *core.Simulation) bool {
 }
 
 func (war *DpsWarrior) tryQueueHsCleave(sim *core.Simulation) {
-	if war.CurrentRage() >= float64(war.Rotation.HsRageThreshold) {
-		if war.Rotation.UseCleave {
-			if war.CanCleave(sim) {
-				war.QueueCleave(sim)
-			}
-		} else {
-			if war.CanHeroicStrike(sim) {
-				war.QueueHeroicStrike(sim)
-			}
-		}
+	if war.ShouldQueueHSOrCleave(sim) {
+		war.QueueHSOrCleave(sim)
 	}
 }
